feat: extract job publishing logic into a reusable PublishJobButton component
